Updating eslint-config-airbnb and eslint-config-prettier to latest. (#18)
This addresses a compatibility issue that arises when a dependency has a peer-dep of a newer version of eslint-plugin-jsx-a11y. Our out-of-date eslint-config-airbnb referenced rules that were removed in eslint-plugin-jsx-a11y@4.*.

Under normal dependencies, this is fine. You can hard lock the version, but create-react-app references eslint-config-react-app which itself requires jsx-a11y 5.x as a peer-dependency. Peer dependencies cannot be meaningfully version locked.

https://nodejs.org/en/blog/npm/peer-dependencies/